Added rotatable camera for character, added debug to take teddy out of death animation, added functionality to change animations for the attacks based on weapons
added rotatable camera, this allows the player to turn the camera around to see the front of the teddy when not moving. Fiddled with star light to make it a launch pad to send teddy from the bed to the dresser. Added R key function to allow debug tester to take teddy out of death animation (teddy can still move while dead i need to find fix for this)  thanks sam for adding weapon changabilty as i can now change weapon type animation