Fixed dodge mechanic, need more dodge animations
i have fixed the dodge mechanic so that the player can roll out of attacks fairly quickly, there is a variable that can be changed by selecting the character in the viewport or going into the character blueprint and selecting the character component. More animations are required as the character bugs out when rolling into the directions between left right forward and backward.